fix gpg verification on nodejs-lts
turns out you need to download a sig/asc file as that has special
behaviour in makepkg. you also need to import the keys from the
keyserver before that step will work. guess this is why antergos did
this.

it worked yesterday for me, but that was because it actually skipped the
verification step (because the .asc wasn't there)